third_party/WebKit/LayoutTests/bluetooth/requestDevice/single-filter-single-service.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setUpHealthThermometerAndHeartRateDevices()
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => assert_equals(device.name, 'Health Thermometer'));
}, 'Simple filter selects matching device.');
</script>
